Fix build errors caused by the (hypothetical) conversion of mlt_position from an int to a float, preserving original behavior.
git-svn-id: https://mlt.svn.sourceforge.net/svnroot/mlt/trunk/mlt@800 d19143bc-622f-0410-bfdd-b5b2a6649095

+++ src/framework/mlt_producer.c
@@ -260,7 +260,7 @@ int mlt_producer_seek( mlt_producer this, mlt_position position )
 	}
 	else if ( use_points && !strcmp( eof, "loop" ) && position >= mlt_producer_get_playtime( this ) )
 	{
-		position = position % mlt_producer_get_playtime( this );
+		position = (int)position % (int)mlt_producer_get_playtime( this );
 	}
 
 	// Set the position
